+++ CHANGES.rst
@@ -48,6 +48,10 @@ New Features
   - Table.read and Table.write now support reading and writing of FITS tables
     via the unified reading/writing interface [#591].
 
+  - The 'units' and 'dtypes' attributes and keyword arguments in Column,
+    MaskedColumn, Row, and Table are now deprecated in favor of the
+    single-tense 'unit' and 'dtype' [#1174].
+
 - :ref:`astropy.vo <astropy_vo>`
 
   - New package added to support Virtual Observatory Simple Cone Search query
@@ -108,6 +112,13 @@ API Changes
     ``copy`` can be used to indicated whether the input data should be copied
     (``copy=True``, default) or used by reference (``copy=False``) in the
     output masked array.  [#1083]
+
+- ``astropy.table``
+
+  - The plural 'units' and 'dtypes' have been switched to 'unit' and 'dtype'
+    where appropriate. The original attributes are still present in this version
+    as deprecated attributes, but will be removed in the next version [#1174].
+
 
 Bug Fixes
 ^^^^^^^^^^